+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## February 22, 2021
+
+### Fix Middleware's bug - Circular import - ([PR #48](https://github.com/ucgmsim/seistech_psha_frontend/pull/48)
+
+- Change the structure a little bit to fix the circular import.
+
 ## Februray 18, 2021
 
 ### GMS Feedback Update Part 1 - ([PR #45](https://github.com/ucgmsim/seistech_psha_frontend/pull/45))